Add error handling to convert script

+++ conv.sh
@@ -26,6 +26,10 @@ function addr2func() {
       done
     } | awk 'NF'
   ))
+  if [[ "${#output[@]}" == "0" ]]; then
+    echo 1>&2 "objdump filter ends with 0 result"
+    return 1
+  fi
   if [[ -p /dev/stdin ]]; then
     perl -ne 'BEGIN { @src_pts=(); @dst_pts=(); $pts_len=$#ARGV / 2; while($#ARGV > 0){ push(@src_pts, shift); push(@dst_pts, shift); }; } for ($i = 0; $i < $pts_len; $i++){ s/@src_pts[$i]/@dst_pts[$i]/; } print $_' "${output[@]}"
   else
@@ -50,6 +54,7 @@ function main() {
 
   if [[ $# -ge 1 ]]; then
     if [[ $1 == '-offset-1' ]]; then
+      echo 1>&2 "[set -offset-1 mode]"
       offset_minux_one_flag=1
       shift
     fi
@@ -76,8 +81,20 @@ function main() {
     cat ./iftracer.out.* | sed -E -e 's/1$/0/' -e 's/5$/4/' -e 's/9$/8/' -e 's/d$/c/' >./iftracer-tmp.out
   fi
 
+  if [[ ! -f "$elf_filepath" ]]; then
+    echo 1>&2 "not found $elf_filepath"
+    return 1
+  fi
+
   echo 1>&2 "[start addr2func]"
   cat ./iftracer-tmp.out | addr2func "$elf_filepath" $(cat ./iftracer-tmp.out | grep -E -o '0[xX][0-9a-fA-F]+' | tr -d '[]' | sort | uniq | tr '\n' ' ') >./iftracer-human.out
+
+  # Is this empty file?
+  if [[ ! -s ./iftracer-human.out ]]; then
+    echo 1>&2 "Maybe you should add -offset-1 option"
+    help
+    return 1
+  fi
 
   echo 1>&2 "[start convert to trace json]"
   {
@@ -120,7 +137,6 @@ EOF
     cat <<EOF
   ],
   "displayTimeUnit": "ns",
-  "systemTraceEvents": "SystemTraceData",
   "otherData": {
     "version": "My Application v1.0"
   }
